[CACHE][CI] Re-organize PDO and logger initialization in cache tests
