+++ tests/TestSuite/Pimcore/Cache/Pool/PdoMysqlCacheItemPoolTest.php
@@ -3,27 +3,9 @@
 namespace TestSuite\Pimcore\Cache\Adapter;
 
 use Cache\IntegrationTests\CachePoolTest;
-use Psr\Cache\CacheItemPoolInterface;
-use Test\Cache\Traits\PdoMysqlCacheItemPoolTrait;
+use TestSuite\Pimcore\Cache\Pool\Traits\PdoMysqlCacheItemPoolTestTrait;
 
 class PdoMysqlCacheItemPoolTest extends CachePoolTest
 {
-    use PdoMysqlCacheItemPoolTrait;
-
-    public static function setUpBeforeClass()
-    {
-        parent::setUpBeforeClass();
-        static::fetchPdo();
-    }
-
-    /**
-     * @return CacheItemPoolInterface that is used in the tests
-     */
-    public function createCachePool()
-    {
-        $itemPool = $this->createPdoItemPool();
-        $itemPool->clear();
-
-        return $itemPool;
-    }
+    use PdoMysqlCacheItemPoolTestTrait;
 }
